fixed issue with back button appearing multiple times at once
