--- Views/TestViews/Test3View.swift
@@ -1,19 +1,0 @@
-//
-//  Test3View.swift
-//  EDUPlate
-//
-//  Created by Ruby Nuss on 11/18/24.
-//
-
-import SwiftUI
-
-struct Test3View: View {
-    @ObservedObject var userProgress: UserProgress
-    var body: some View {
-        TestView(category: "French Terms", userProgress: userProgress)
-    }
-}
-
-#Preview {
-    Test3View(userProgress: UserProgress())
-}
