+++ f5.json
@@ -68,5 +68,6 @@
 {"category": "Ingredients", "question": "Chives", "answer": "has a mild, onion-like flavor, with a hint of garlic", "fake1": "a licorice-like flavor popular in candies", "fake2": "has an anise-like flavor but is more aromatic, sweeter and less pungent", "fake3": "the rich, brown, dried, unopened flower buds of Syzygium aromaticum, an evergreen tree in the myrtle family."} ,
 {"category": "Ingredients", "question": "Cilantro", "answer": "has a fragrant mix of parsley and citrus", "fake1": "has little aroma, but it is extremely hot to taste.", "fake2": "used in spice cookies and cakes. strong, pungent, and sweet.", "fake3": "has brilliant yellow color, and has been used as a dye, medicine, and flavoring since 600 BC."} ,
 {"category": "Ingredients", "question": "Cream of Tarter", "answer": "used to stabilize egg white foams. It is also a major ingredient in baking powder.", "fake1": "used in Indian curries, gin and sausages.", "fake2": "A necessary ingredient of curry powder", "fake3": "used as a thickening agent for sauces, fruit pie fillings and glazes"} ,
-{"category": "Ingredients", "question": "Fennel", "answer": "has an anise-like flavor but is more aromatic, sweeter and less pungent", "fake1": "the rich, brown, dried, unopened flower buds of Syzygium aromaticum, an evergreen tree in the myrtle family.", "fake2": "has brilliant yellow color, and has been used as a dye, medicine, and flavoring since 600 BC.", "fake3": "Add it to stews and marinades for a Mediterranean flavor."}
+{"category": "Ingredients", "question": "Fennel", "answer": "has an anise-like flavor but is more aromatic, sweeter and less pungent", "fake1": "the rich, brown, dried, unopened flower buds of Syzygium aromaticum, an evergreen tree in the myrtle family.", "fake2": "has brilliant yellow color, and has been used as a dye, medicine, and flavoring since 600 BC.", "fake3": "Add it to stews and marinades for a Mediterranean flavor."},
+{"category": "Basic Cooking Techniques", "question": "The correct description for boiling is", "answer": "Large bubbles breaking at the surface 100 degrees celsius", "fake1": "small bubbles breaking at the surface 98 degrees celsius", "fake2": "Water reaching 98 degrees celsius", "fake3": "Frying in oil"}
 ]
